app12/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-multipart==0.0.6
nltk==3.8.1 